Mark Rober debugging
It just lets the blueLED glow ... it stays in the last state and does not loop through the other cases (first to ifs) even though the variable x is changing. :( When i restart the program first it gets through the stages, i also managed to loop through all the combinations..... so that it would cicle .... Do you have any ideas?